SCE-615: Local executor doesn't cleanup output files when Execute() fails (#315)
* SCE-615: Local executor doesn't cleanup output files when Execute()
fails

* Typo fix

* Following fixup

* Better parent directoru detection

* Apply @skonefal review

* Remove unused import

* Apply @wlodek's review

* Fix tests

* Remove unused import

* Logic Fix

+++ pkg/executor/local.go
@@ -3,7 +3,7 @@ package executor
 import (
 	"os"
 	"os/exec"
-	"path"
+	"path/filepath"
 	"syscall"
 	"time"
 
@@ -46,6 +46,7 @@ func (l Local) Execute(command string) (TaskHandle, error) {
 
 	stdoutFile, stderrFile, err := createExecutorOutputFiles(command, "local")
 	if err != nil {
+		eraseOutput(stdoutFile)
 		return nil, errors.Wrapf(err, "createExecutorOutputFiles for command %q failed", command)
 	}
 
@@ -57,6 +58,7 @@ func (l Local) Execute(command string) (TaskHandle, error) {
 
 	err = cmd.Start()
 	if err != nil {
+		eraseOutput(stdoutFile)
 		return nil, errors.Wrapf(err, "command %q start failed", command)
 	}
 
@@ -242,9 +244,16 @@ func (taskHandle *localTaskHandle) Clean() error {
 
 // EraseOutput removes task's stdout & stderr files.
 func (taskHandle *localTaskHandle) EraseOutput() error {
-	outputDir, _ := path.Split(taskHandle.stdoutFile.Name())
+	return eraseOutput(taskHandle.stdoutFile)
+}
 
+// eraseOutput requires only one file from stdout and stderr files. It's looking for parent dir and remove it.
+func eraseOutput(outputFile *os.File) error {
 	// Remove temporary directory created for stdout and stderr.
+	if outputFile == nil {
+		return nil
+	}
+	outputDir := filepath.Dir(outputFile.Name())
 	if err := os.RemoveAll(outputDir); err != nil {
 		return errors.Wrapf(err, "os.RemoveAll of directory %q failed", outputDir)
 	}
